// temp/cms/models/com_ever_edu_cms_scorm_dto_rte_ScoContentSettingResDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { com_ever_edu_cms_scorm_dto_rte_ScoCmiObjectInfoDto } from './com_ever_edu_cms_scorm_dto_rte_ScoCmiObjectInfoDto';
export type com_ever_edu_cms_scorm_dto_rte_ScoContentSettingResDto = {
    sequenceId?: number;
    courseId?: number;
    curriculumId?: number;
    contentId?: number;
    orgnId?: number;
    scoId?: string;
    itemURL?: string;
    previous?: boolean;
    next?: boolean;
    objectInfo?: com_ever_edu_cms_scorm_dto_rte_ScoCmiObjectInfoDto;
};